Adding ops_repo.source attribute
