+++ Dockerfile-core
@@ -27,6 +27,4 @@ ONBUILD WORKDIR /opt/bitops
 # ONBUILD COPY scripts/ ./scripts has proven quite useful for rapid testing. Please keep in while testing. 
 ONBUILD COPY scripts/ ./scripts
 ONBUILD COPY bitops.config.yaml .
-#ONBUILD COPY cloudformation.schema.yaml .
-#ONBUILD COPY cloudformation.config.yaml .
 ONBUILD RUN python3 scripts/plugins.py install
