+++ scripts/plugins.py
@@ -1,19 +1,17 @@
 import sys
 import os
 
-from plugins.settings import BITOPS_logging_level
+import plugins.settings
+
 from plugins.logging import logger
 from plugins.deploy_plugins import Deploy_Plugins
 from plugins.install_plugins import Install_Plugins
-from plugins.utilties import Get_Config_List
 
 if __name__ == "__main__":
     try:
         mode = sys.argv[1]
     except IndexError:
         mode = None
-
-    logger.debug(os.environ)
 
     if mode == "deploy":
         Deploy_Plugins()
